chore: exportable parse()
