+++ scripts/src/types/github-issue.ts
@@ -2,3 +2,7 @@ export type Label = {
   name: string;
   color: string;
 };
+
+export interface Payload {
+  [key: string]: string;
+}
