+++ scripts/src/utils/github.ts
@@ -1,4 +1,4 @@
-import { Label } from "../types";
+import { Label, Payload } from "../types";
 
 export const formatKey = (key: string): string =>
   (key || "").replace("### ", "").replace(" ", "_").toLowerCase();
@@ -6,30 +6,32 @@ export const formatKey = (key: string): string =>
 export const isKey = (key: string) =>
   (key || "").includes("### ") && !(key || "").includes("Code of Conduct");
 
-export const getType = (labels: Array<Label>): string | null => {
+export const getType = (labels: Array<Label>): string => {
   if (labels.length != 2) {
-    return null;
+    throw new Error("Labels are empty");
   }
 
   const isNew = labels.find((l) => l.name === "new");
   const type = labels.find((l) => l.name !== "new");
 
-  if (!isNew && !type) {
-    return null;
+  if (!isNew || !type || !type.name) {
+    throw new Error("An error error occurred ");
   }
 
   return type?.name;
 };
 
 export const parseAttributes = (body: string) => {
+  // TODO: make the separator a function  parameter
   const lines = body.split("\\n\\n");
-  const payload = {
+  const payload: Payload = {
     createdAt: new Date().toISOString(),
   };
   for (let index = 0; index < lines.length; index++) {
     const line = lines[index];
+    const key: string = formatKey(line);
     if (isKey(line)) {
-      payload[formatKey(line)] = lines?.[index + 1];
+      payload[key] = lines?.[index + 1];
     }
   }
   return payload;
